requirements.txt
fastapi
vellox
google-genai
httpx
python-dotenv